requirements.txt
pathvalidate==2.0.1
requests==2.22.0
lxml==4.4.2
beautifulsoup4==4.8.2
Pillow==7.0.0
